refactor: remove dead production code
